Yes, i did it, calc_word_value can calculate only one word, max_word_value can calculate dict.
